# LSM: Unified {table_memory, prefetch + cache}, Scopes, Cache Map