The primary purpose behind this PR is to optimize, and enable future optimizations for the CPU-bound parts of compaction, specifically including the ability to pace the previously one-off mutable table sort.

This PR unifies our `TableMutable` and `TableImmutable` backing implementations into a single `TableMemory`, hoists our Tree cache to rather be a Groove cache (keyed by ID instead of timestamp, and cutting out coherency complexity) as well as unifies the new Groove cache with the prefetch objects.

Additionally, there's a new data structure - `CacheMap` - a union between our `SetAssociativeCache` and a `HashMap`. There's also a new feature, for state machines, Scopes, enabled by the above changes that make it much simpler to write code that involves rollbacks.

# Show-me-the-numbers
|                                                      | main tps | this pr tps | main p100 | this pr p100 |
|------------------------------------------------------|----------|-------------|-----------|--------------|
| benchmark as is (10k acc), nvme                      | 121k     | 150k        | 5199ms    | 4286ms       |
| benchmark as is (10k acc), tmpfs                     | 230k     | 316k        | 1778ms    | 1616ms       |
| benchmark 10 accounts, nvme                          | 132k     | 163k        | 4932ms    | 4165ms       |
| benchmark 10 accounts, tmpfs                         | 266k     | 377k        | 1532ms    | 1740ms       |
| benchmark 100k acc, nvme                             | 117k     | 145k        | 5531ms    | 4516ms       |
| benchmark 100k acc, tmpfs                            | 215k     | 296k        | 1627ms    | 1590ms       |
| (bonus) benchmark 10 accounts, sequential ids, tmpfs | 335k     | 488k        | 1704ms    | 952ms        |

## Cache Map
The CacheMap is backed by two data structures: our `SetAssociativeCache` providing the Cache layer, and a `HashMap` providing the Stash layer. Functionally, it behaves as a cache where a value can be _guaranteed_ to be present for a certain period of time. 

This allows use cases beyond caching; namely where a value is required to be present if it has been inserted. A set associative cache, even if large enough, could suffer from collisions and not maintain this property. In this PR, it's the backing for the new combined Groove prefetch + cache. 

Additionally, it has scope support and a fuzzer.

## Unified, log based `TableMemory`
Previously, there were two different implementations providing `TableMutable` and `TableImmutable`. We now unify the implementation (there is still a separate `table_mutable` and `table_immutable`!). This allows for pointer swapping when swapping the mutable and immutable tables.

Additionally, the implementation for both is now array based rather than hash map based for the mutable table previously. Inserts are extremely cheap, but removals have to be handled by tombstones. This does regress an optimization we had, but more on that later.

Currently, `table_mutable` is still sorted all in one go before being swapped to `table_immutable`. Work to make this incremental, using some k-way strategy will be done after this. `TableMemory` also implements scope support.

## Unified Groove level object cache + prefetch
The Tree based cache has been removed, and replaced with a Groove level cache that's keyed by ID. Since we no longer have a hash map to do cheap lookups on data we inserted into `table_mutable` or `table_immutable` (which used to do a binary search), we actually rely on the Groove cache to be our working copy of in memory data.

The data in the Groove cache is used to serve point lookups, and may be duplicated in `TableMemory`. Future range queries will likely use the data there, and ignore the Groove cache.

This unlocks a neat optimization: our object cache is now also our prefetch objects! If the entry is already in the cache, we don't need to copy it from our cache, it's already there.

## Scopes
Previously, the state machine needed to define custom `_rollback` operations that would perform the inverse operations to return everything to the state it needed to be in. With both `TableMemory` and `CacheMap` having scope support, it's instead now possible to do:

```
self.scope_open();
const all_things_succeeded = self.do_a_bunch_of_things();
self.scope_close(if all_things_succeeded .persist else .discard);
```

Scopes are designed so that persisting is a cheap operation, while rollbacks pay a small performance hit to restore the state of the Groove cache.

# Regressions
The secondary index optimization at https://github.com/tigerbeetle/tigerbeetle/pull/337/files is no more, because our TableMemory log can't absorb duplicates in the same way a HashMap would. Additionally, things that perform multiple updates (eg, modifying the same AccountMutable 10 times) are no longer absorbed either, and have to be fully sorted first before being filtered out. 

# Future work
* Making the mutable table sort fully incremental (exploring KWayMerge strategies).
* Check if making TableMemory.put not update the `sorted` property on each put, but rather in a hot loop per batch will be faster.
* Improving the secondary index optimization above.